# Efficient implementation for BitSet minAfter / maxBefore