Currently `BitSet` is using implementations of `minAfter`/`maxBefore` inherited from `SortedSet`. These implementations are very inefficient for `BitSet`:
- allocate a clone of `BitSet` and erase bits outside of the range (`rangeFrom`/`rangeUntil`)
- allocate and traverse first `Iterator` for `isEmpty` check
- allocate and traverse another `Iterator` for `head` / `last`

This PR implements dedicated overrides of `minAfter`/`maxBefore` for `BitSet`. No object allocations are performed, and a minimal traversal of `BitSet` words is done.

Current implementation:
```
[info] Benchmark                               (spacing)  Mode  Cnt     Score    Error  Units
[info] BitSetBoundedMinMaxBenchmark.maxBefore          0  avgt    6  2058.881 ± 23.838  ns/op
[info] BitSetBoundedMinMaxBenchmark.maxBefore          8  avgt    6   345.129 ±  4.539  ns/op
[info] BitSetBoundedMinMaxBenchmark.maxBefore        255  avgt    6    87.353 ±  0.714  ns/op
[info] BitSetBoundedMinMaxBenchmark.minAfter           0  avgt    6    64.473 ±  1.092  ns/op
[info] BitSetBoundedMinMaxBenchmark.minAfter           8  avgt    6    62.591 ±  0.859  ns/op
[info] BitSetBoundedMinMaxBenchmark.minAfter         255  avgt    6    62.375 ±  0.426  ns/op
```

This PR:
```
[info] Benchmark                               (spacing)  Mode  Cnt   Score   Error  Units
[info] BitSetBoundedMinMaxBenchmark.maxBefore          0  avgt    6  13.595 ± 0.148  ns/op
[info] BitSetBoundedMinMaxBenchmark.maxBefore          8  avgt    6  13.725 ± 0.344  ns/op
[info] BitSetBoundedMinMaxBenchmark.maxBefore        255  avgt    6  15.008 ± 0.535  ns/op
[info] BitSetBoundedMinMaxBenchmark.minAfter           0  avgt    6  13.334 ± 0.134  ns/op
[info] BitSetBoundedMinMaxBenchmark.minAfter           8  avgt    6  13.270 ± 0.155  ns/op
[info] BitSetBoundedMinMaxBenchmark.minAfter         255  avgt    6  16.770 ± 0.054  ns/op
```

Benchmark code:
```scala
package scala.collection.mutable

import java.util.concurrent.TimeUnit

import org.openjdk.jmh.annotations._

import scala.collection.mutable

@BenchmarkMode(Array(Mode.AverageTime))
@Fork(1)
@Threads(1)
@Warmup(iterations = 6)
@Measurement(iterations = 6)
@OutputTimeUnit(TimeUnit.NANOSECONDS)
@State(Scope.Benchmark)
class BitSetBoundedMinMaxBenchmark {

  @Param(Array("0", "8", "255"))
  var spacing: Int = _

  var bs: mutable.BitSet = _
  var middle: Int = _

  @Setup(Level.Iteration) def initializeRange(): Unit = {
    bs = mutable.BitSet(0 until 1000 by (spacing + 1): _*)
    middle = bs.max / 2
  }

  @Benchmark def minAfter = bs.minAfter(middle)

  @Benchmark def maxBefore = bs.maxBefore(middle)

}
```