@linasm 2020-01-06
Failed with:
```
mimaReportBinaryIssues:
[error]  * method maxBefore(Int)scala.Option in interface scala.collection.BitSetOps's type is 
different in other version, where it is (java.lang.Object)scala.Option instead of (Int)scala.Option
(...)
```
Is there a way to work around this?