@diesalbla 2020-01-06
@linasm That MIMA error comes because `maxBefore` is inherited from `Set`, where it is generic. A generic method, in Bytecode, takes an `Object`. In the case of a `BitSet`, that is a _boxed_ integer.