@diesalbla 2020-01-06
Here is a suggestion: can you define in `Set` a protected, inner type alias for the `A` type parameter? Then, perhaps you can define `maxBefore` on that type. In source it may infer that 

```Scala
trait Set[A] {
  protected type Elem = A
}

class BitSet extends Set[Int] {
   override def maxBefore(a: Elem): Option[Elem]
}
```

I am not sure if this would work, but perhaps it's worth a try..