@linasm 2020-01-07
Thanks for the suggestion, @diesalbla . Unfortunately, looks like it did not help.
I'm trying to look into the encoding of methods, before the change:
```java
   public Option maxBefore(final Object key) {
      return SortedSetOps.maxBefore$(this, key);
   }
``` 
After the change:
```java
   public Option maxBefore(final int key) {
      return BitSetOps.maxBefore$(this, key);
   }


   // $FF: synthetic method
   // $FF: bridge method
   public Option maxBefore(final Object key) {
      return this.maxBefore(BoxesRunTime.unboxToInt(key));
   }
```

In my very little understanding of the issue, the original boxing method taking `Object` is preserved, and the new method taking a primitive `int` is just an additional overload.

Can someone confirm that this is indeed a binary compatibility violation, and not a false positive?