@sjrd 2020-01-07
I confirm that this is indeed a binary compatibility violation. Forward bincompat, to be precise.

See also https://github.com/scala/bug/issues/11804