@linasm 2020-01-07
@sjrd thanks, mentioning forward compatibility finally made it obvious to me.

Seeing that this change is no go in any nearby future, should I just close this PR,
or can in be somehow labeled and stay around until better days?

IMHO performance improvement is quite substantial.