@szeiger 2020-01-13
Tentatively scheduling for 2.14.0-M1 (which will not exist but is used as a placeholder for things we would want to merge if it wasn't for binary compatibility constraints). We can revisit this if https://github.com/scala/bug/issues/11804 gets implemented.