@lrytz 2020-01-14
If we were to merge this after a fix for https://github.com/scala/bug/issues/11804, the parameter type would be `Object` instead of `Int`, so the compiler would have to add the unbox operations to the implementation. Should work fine, but it can have an unexpected performance penalty in general (here it probably wouldn't matter).