@szeiger 2020-01-14
It's already `Object` now. You could benchmark the difference in the new implementation by upcasting to `SortedMap` in the benchmarking code.