@dwijnand 2020-10-05
"Draft"ing while scala/bug#11804 is still pending.  And even with that it would need its mima exclusions.