@dwijnand 2020-12-04
Going to add this to the list in https://github.com/scala/bug/issues/11804#issuecomment-738879217, and otherwise close this one as we cannot land it without that.